TEST: Move datapath back to util file
